==> notebooks/02_generating_ground_truth_data.ipynb <==
from tqdm.auto import tqdm
import pandas as pd
from dotenv import load_dotenv
import os
from openai import OpenAI
import json

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

df = pd.read_csv('../data/krakow_pois_for_rag.csv')

text_columns = ['name','amenity','leisure','natural','tourism','historic','wiki_summary_en']

openai_client = OpenAI(api_key=OPENAI_API_KEY)


df.shape

There is too many records - I will select all with wiki_descitption and same number of rows w/o description.

df_with_wiki = df[df['wiki_summary_en'] != 'no information']


df_with_wiki.shape

df_without_wiki = df[df['wiki_summary_en'] == 'no information'].sample(df_with_wiki.shape[0], random_state=123)

df_without_wiki.shape

df = pd.concat([df_with_wiki, df_without_wiki])


df.shape

df.to_csv('../data/krakow_pois_selected.csv', index=False,header=True)

documents = df.to_dict(orient='records')

prompt_template = """
Generate a list of realistic and diverse questions that a traveler might ask a virtual travel assistant while visiting Kraków, Poland.

Based on the given attraction record, which is Krakow's POI (Point of interest), create 5 complete, specific questions the user might ask about the attraction. 
Only create questions for categories that contain information different than "No data available".
The questions must:

* Be relevant to the details in the record.
* Be clear and self-contained (not too short).
* Use as few exact words from the record as possible while keeping the meaning.

Attraction record format:

phone : {phone}
cemetery : {cemetery}
emergency : {emergency}
opening_hours : {opening_hours}
website : {website}
pets_allowed : {pets_allowed}
geometry : {geometry}
historic : {historic}
wiki_summary_en : {wiki_summary_en}
postal_code : {postal_code}
toilets : {toilets}
natural : {natural}
description : {description}
visiting_time : {visiting_time}
leisure : {leisure}
tourism : {tourism}
public_transport : {public_transport}
brand : {brand}
alt_name : {alt_name}
amenity : {amenity}
reservation : {reservation}
attraction : {attraction}
highchair : {highchair}
parking : {parking}
swimming_pool : {swimming_pool}
contact_phone : {contact_phone}
community_centre : {community_centre}
addr_street : {addr_street}
contact_twitter : {contact_twitter}
social_facility : {social_facility}
contact_facebook : {contact_facebook}
zoo : {zoo}
email : {email}
wheelchair : {wheelchair}
cuisine : {cuisine}
contact_website : {contact_website}
internet_access : {internet_access}
opening_hours_reception : {opening_hours_reception}
guest_house : {guest_house}
addr_city : {addr_city}
contact_instagram : {contact_instagram}
image : {image}
location : {location}
outdoor_seating : {outdoor_seating}
museum : {museum}
takeaway : {takeaway}
smoking : {smoking}
name : {name}
id : {id} 

Provide the output in parsable JSON without using code blocks:

{{"questions": ["question1", "question2", ..., "question5"]}}

""".strip()

def generate_questions(doc):

    prompt = prompt_template.format(**doc)

    response = openai_client.chat.completions.create(
        model='gpt-4o-mini',
        messages=[{"role": "user", "content": prompt}]
    )

    json_response = response.choices[0].message.content
    return json_response

documents[0]['name']

questions = generate_questions(documents[0])


json.loads(questions)

results = {}

import re

for doc in tqdm(documents):
        doc_id = doc['id']
        if doc_id in results:  # Skip if already processed
            continue


        questions_raw = generate_questions(doc)
        questions_raw = re.sub(r'(\?")\s*([A-Z])', r'?", "\2', questions_raw)
        questions = json.loads(questions_raw)


        results[doc_id] = questions['questions']

final_results = []

for doc_id, questions in results.items():
    for q in questions:
        final_results.append((doc_id, q))

df_results = pd.DataFrame(final_results, columns=['id', 'question'])

df_results.tail(10)

df_results.shape

df_results.to_csv('../data/ground-truth-retrieval.csv', index=False)
